Move chain sync functions to Cardano.Api.LedgerState.ChainSync
